Use `TranscriptRng` for verifier weights (#10)
As noted in #6, verification requires a suitable random number generator, which the target may not be able to provide. The generator is used to produce weights that are applied to each verification equation for efficiency.

This PR makes the verifier deterministic by deriving the weights from a pseudorandom number generator derived from the proof transcript after adding the prover responses to it. This means the prover cannot modify the proof without changing the weights, making it infeasible to produce an invalid proof that passes verification.

Closes #6.

Supersedes #8.

BREAKING CHANGE: Changes the public API by removing the requirement that the verifier supply a random number generator.