Variable-time prover (#19)
Prover operations make some attempt at avoiding timing side-channel attacks through the use of constant-time multiscalar multiplication operations that are used in several places. While this is useful in general, some callers might not need this, and would prefer a speedier prover.

This PR adds a `prove_vartime` function that uses variable-time multiscalar multiplication, which cuts the proving time by about half. A simple refactoring means it and the existing `prove` function are now trivial wrappers to avoid code duplication.

Tests and benchmarks are updated to account for this.