Add transcript composition (#30)
This PR adds transcript composition, which allows for Triptych proofs to be more seamlessly used as part of larger protocols. To do so, the API is modified such that the prover and verifier accept a mutable Merlin transcript. This replaces the existing message binding, which is offloaded to the transcript.

Closes #28.

BREAKING CHANGE: Updates the APIs for `Statement` and `Proof`.