Add `rand` feature and update defaults (#38)
This PR adds a `rand` feature. This enables two new prover functions, `prove_with_rng` and `prove_vartime_with_rng`, that use `OsRng` as the random number generator. This makes it easier to securely generate proofs on targets that support this.

It also updates the default features to include `rand`, `serde`, and `std` to provide more functionality out of the box.