Rename functions (#39)
Recent work in #38 adds prover functionality that uses `OsRng` instead of requiring the caller to supply a random number generator. The intent was to maintain API compatibility, but this resulted in function names that are opposite what you might expect.

This PR renames the prover functions. The `prove` and `prove_vartime` functions require the `rand` feature and use `OsRng`. The `prove_with_rng` and `prove_with_rng_vartime` functions require the caller to supply a random number generator.

BREAKING CHANGE: This changes the API by renaming functions.