Custom serialization (#44)
Serialization using `serde` is suboptimal. Depending on the encoding used, serialization may not be canonical, and is likely to be inefficient due to length encoding.

This PR adds custom serialization and deserialization that is efficient and canonical.

Because this is easy to mess up, it also adds a basic fuzzer. The fuzzer attempts to deserialize byte slices; deserialization should either fail or subsequently serialize canonically.

Note that `serde`-based serialization is still supported.

Closes #43.