Check witness consistency in constant time (#49)
This PR adds a constant-time check of witness consistency to the prover. It does this by iterating over the entire input set and doing a conditional assignment using the secret index. This should remove any timing effects caused by vector lookups.

Note that this is only used in the constant-time prover functions, so variable-time proving is unaffected.