Allow input set padding (#57)
This PR adds optional input set padding, which allows you to specify an input set that is too short for the given parameters. It does this by repeating the last element of the input set until the result is the correct length. Because not all use cases may be able to do this safely, this functionality is added via a new `InputSet::new_with_padding` function.

Closes #56.