Rename generator (#91)
Parallel proofs support the use of separate generators for verification and auxiliary verification keys. Currently, the notation `H` is used for the latter, which may be confusing depending on use case by conflicting with other notation (like Pedersen commitments).

This PR renames the generator to `G1`, which is also more in line with how input sets are defined. Note that this is also reflected in proof transcripts.

BREAKING CHANGE: Changes the public API for parallel proofs. Existing proofs will not verify.